Move some widgets to the console, being the core plugin of spyder
