+++ spyder/api/widgets/dependencies.py
