+++ spyder/api/widgets/dock.py
