+++ spyder/plugins/console/widgets/about.py
