+++ spyder/plugins/console/widgets/reporterror.py
